+++ src/constants/ExcelEmulator/delays.ts
@@ -1,0 +1,1 @@
+export const helpMessageDelay = 10000; // toastAutoCloseTimeout + 2000;