+++ src/pages/ExcelEmulatorPage.tsx
@@ -1,14 +1,17 @@
 import { ExcelEmulatorScreen } from '@/components/ExcelEmulator/ExcelEmulatorScreen';
+import { FireworksContextProvider } from '@/contexts/FireworksContext';
 import { ProgressContextProvider } from '@/contexts/ProgressContext';
 import { SelectionContextProvider } from '@/contexts/SelectionContext';
 
 export function ExcelEmulatorPage() {
   return (
-    <ProgressContextProvider>
-      <SelectionContextProvider>
-        {/* Nested components */}
-        <ExcelEmulatorScreen />
-      </SelectionContextProvider>
-    </ProgressContextProvider>
+    <FireworksContextProvider>
+      <ProgressContextProvider>
+        <SelectionContextProvider>
+          {/* Nested components */}
+          <ExcelEmulatorScreen />
+        </SelectionContextProvider>
+      </ProgressContextProvider>
+    </FireworksContextProvider>
   );
 }